LMN-1159 Add price row variant
