+++ backpack-compose/src/main/kotlin/net/skyscanner/backpack/compose/price/BpkPrice.kt
@@ -22,10 +22,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import net.skyscanner.backpack.compose.price.internal.BpkPriceAlignEnd
 import net.skyscanner.backpack.compose.price.internal.BpkPriceAlignStart
+import net.skyscanner.backpack.compose.price.internal.BpkPriceRow
 
 enum class BpkPriceAlign {
     Start,
     End,
+    Row,
 }
 
 enum class BpkPriceSize {
@@ -55,8 +57,20 @@ fun BpkPrice(
                 size = size,
             )
         }
+
         BpkPriceAlign.End -> {
             BpkPriceAlignEnd(
+                price = price,
+                modifier = modifier,
+                leadingText = leadingText,
+                previousPrice = previousPrice,
+                trailingText = trailingText,
+                size = size,
+            )
+        }
+
+        BpkPriceAlign.Row -> {
+            BpkPriceRow(
                 price = price,
                 modifier = modifier,
                 leadingText = leadingText,
